+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
 - Adjusted `Endpoint::body` to return an `Option` on success
 - Adjusted `Endpoint::body` and `Endpoint::parse` to return
   `Endpoint::ConversionError` on failure
+- Adjusted `Endpoint::query` return a `Result`
 - Require `std::error::Error` instead of `std::fmt::Display` for
   `Endpoint::ApiError` type
 - Removed `DeserializeOwned` requirement from `Endpoint::Output` and